--- frontend/src/components/BaseErrorBoundary.tsx
@@ -1,70 +1,0 @@
-import { Component, ErrorInfo, ReactNode } from 'react';
-
-interface Props {
-  children: ReactNode;
-  fallback?: ReactNode;
-  onError?: (error: Error, errorInfo: ErrorInfo) => void;
-}
-
-interface State {
-  hasError: boolean;
-  error: Error | null;
-  errorInfo: ErrorInfo | null;
-}
-
-class BaseErrorBoundary extends Component<Props, State> {
-  public state: State = {
-    hasError: false,
-    error: null,
-    errorInfo: null
-  };
-
-  public static getDerivedStateFromError(error: Error): Partial<State> {
-    return { hasError: true, error };
-  }
-
-  public componentDidCatch(error: Error, errorInfo: ErrorInfo) {
-    this.setState({ errorInfo });
-
-    this.props.onError?.(error, errorInfo);
-
-    if (import.meta.env.DEV) {
-      console.error('Uncaught error:', error);
-      console.error('Error info:', errorInfo);
-    }
-  }
-
-  public render() {
-    if (this.state.hasError) {
-      if (this.props.fallback) {
-        return this.props.fallback;
-      }
-
-      return (
-        <div className="fixed w-screen h-screen z-50 flex items-center justify-center bg-gray-100/80">
-          <div className="p-8 bg-white rounded-xl shadow-lg max-w-md w-full">
-            <h2 className="text-2xl font-bold text-red-600 mb-4">Something went wrong</h2>
-            <p className="text-gray-600 mb-4">
-              {this.state.error?.message || 'An unexpected error occurred'}
-            </p>
-            {import.meta.env.DEV && (
-              <pre className="bg-gray-100 p-4 rounded mb-4 text-sm overflow-auto max-h-40">
-                {this.state.errorInfo?.componentStack}
-              </pre>
-            )}
-            <button
-              onClick={() => window.location.reload()}
-              className="bg-blue-500 text-white px-4 py-2 rounded hover:bg-blue-600 transition-colors"
-            >
-              Refresh Page
-            </button>
-          </div>
-        </div>
-      );
-    }
-
-    return this.props.children;
-  }
-}
-
-export default BaseErrorBoundary;
