+++ frontend/src/features/Chat/components/ChatListItem.tsx
@@ -1,8 +1,8 @@
 import { useEffect, useState } from 'react';
-import { useAuthStore } from '../../../../../store/useAuthStore';
+import { useAuthStore } from '../../../store/useAuthStore';
 import { ChatRoom } from '../types/chatTypes';
-import { profileService } from '../../../services/profileService';
-import { formatRelativeTime } from '../../../../../utils/dateUtils';
+import { profileService } from '../../shared/services/profileService';
+import { formatRelativeTime } from '../../../utils/dateUtils'; 
 import { chatService } from '../services/chatService';
 
 interface ChatListItemProps {
