+++ frontend/src/features/shared/components/Sidebar.tsx
