+++ frontend/src/features/shared/components/SidebarItem.tsx
