+++ frontend/src/pages/Main/Home.tsx
@@ -1,8 +1,8 @@
-import Sidebar from '../../features/Main/components/Sidebar';
-import HomeProfile from '../../features/Main/section/MainHome/components/HomeProfile';
-import CurrentLocationMap from '../../features/Main/section/MainHome/components/CurrentLocationMap';
-import TravelerCardList from '../../features/Main/section/MainHome/components/TravelerCardList';
-import { useUserLocation } from '../../features/Main/section/MainHome/hooks/useUserLocation';
+import Sidebar from '../../features/shared/components/Sidebar';
+import HomeProfile from '../../features/MainHome/components/HomeProfile';
+import CurrentLocationMap from '../../features/MainHome/components/CurrentLocationMap';
+import TravelerCardList from '../../features/MainHome/components/TravelerCardList';
+import { useUserLocation } from '../../features/MainHome/hooks/useUserLocation';
 import DataFetchErrorBoundary from '../../components/ErrorBoundary/DataFetchErrorBoundary';
 
 const Home = () => {
