--- frontend/src/store/useLoadingStore.ts
@@ -1,11 +1,0 @@
-import { create } from 'zustand';
-
-interface LoadingState {
-  isLoading: boolean;
-  setLoading: (loading: boolean) => void;
-}
-
-export const useLoadingStore = create<LoadingState>((set) => ({
-  isLoading: false,
-  setLoading: (isLoading) => set({ isLoading })
-}));
